+++ flutter_modular/example/lib/app/modules/shopping/pages/page2/page2_bloc.dart
@@ -1,7 +1,0 @@
-import 'package:flutter_modular/flutter_modular.dart';
-
-class Page2Bloc extends Disposable {
-  //dispose will be called automatically by closing its streams
-  @override
-  void dispose() {}
-}