+++ flutter_modular/lib/src/presenters/main_module.dart
@@ -1,0 +1,7 @@
+import 'package:flutter/widgets.dart';
+
+import '../core/interfaces/child_module.dart';
+
+abstract class MainModule extends ChildModule {
+  late final Widget bootstrap;
+}