+++ flutter_modular/test/src/core/models/bind_test.dart
@@ -1,0 +1,18 @@
+import 'package:flutter_modular/src/core/models/bind.dart';
+import 'package:flutter_test/flutter_test.dart';
+
+main() {
+  test('check type', () {
+    final bind = Bind.instance('teste');
+    expect(bind, isA<Bind<String>>());
+  });
+
+  test('check type in list', () {
+    final binds = [
+      Bind.instance('teste'),
+      Bind.instance(true),
+    ];
+
+    //  expect(bind, isA<Bind<String>>());
+  });
+}